Made Editor Script
Editor Script to spawn Ammo and Med boxes as AI could not gather a list on start for either. Also changed some values within the AI on both Soldier and SoldierTest2 prefabs. BUG ONLY SOLDIER 2 COMES ONTO THE OVERLAY THOUGH WITHIN THE CODE THE AMOUNT OF AGENTS IS THE AMOUNT THAT IS WITHIN THE ENVIRONMENT YET ONLY SOLDIER 2 SHOWS