Updates
Updated the AILogic script to were it was before the fatal error. Soldier 1 wants health and ammo yet stil moves towards the enemy, Soldier 2 does the same as 1 but instead interrupts Find Enemy, Soldier 3 just has Get Health always runnings so need to adjust weightings and curves for the AI to produce something that works in a BR game mode. Started working on more UI just by setting up multiples camera's Main to move around the map, then when selecting an AI agent it goes to the Camera attached to the AI agent and then deselect goes back to Main camera