Last Changes
Caused AI to no longer move so made a new branch from this one before commiting these changes. This is to see if the branch i am working on with the last push is the working version of the AI moving.

Actual changes are Ammo State to Use in the AILogic check for GetAmmo Action. The same as the Health Component Health State. Changed the AILogic script so that in each function for the action it increases or decrease the Utility Values. Then when we Determine What to Do and check if the delegate handle is equal to the action then execute the code in there for attacking and choosing the best enemy or med/ammo box for Get Health and Get Ammo.