Update
AI agent 3 and 1 attack enemies, they don't choose the best position but can do much about it, problem is they have to get to the destination before they change direction again, can't change that due to it being the Nav Agent. Both Agents have animations running, problem is when an agent dies they don't remove from the UI and are still in the environment so the user can still access them and the data.